LMW-185 Move tests to the right place
